requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
ccxt==3.0.61
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
pycares==4.3.0
pycparser==2.21
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
